more changing of int to long

+++ Problem4.java
@@ -32,12 +32,12 @@ public class Problem4 {
             }
         }
         //printArray(foods); // debug
-        int r = 0;
-        int c = 0;
+        long r = 0;
+        long c = 0;
         boolean horiz = true;
-        int count = 0;
-        int j = 0;
-        int i= 0;
+        long count = 0;
+        long j = 0;
+        long i= 0;
         //System.out.println("debug: count =" + count + " r = " + r + " c = " + c + " i = " + i + " j = " + j +
                 //"print index = " + printIndicies[j] + " instruction = " + instructions[i]);
         for (i = 0; i < instructions.length ; i++) {
